Fix the missing derivative member check
Close #8568.

The root cause of this issue is that when the struct is indirectly
inherited from IDifferentiable type, we will not check the reference
of the DerivativeMember attribute. This PR fixes this issue by checking
whether a struct is conformed to IDifferentiable correctly, and it is
we will always check the DerivativeMember attribute before the type
conformance check of the struct.