Add an __undefined literal expression
This change adds a new form of literal expression, `__undefined`, which can be used as the initial value for variables to indicate that the compiler should emit no code to compute or store an initial value (beyond whatever might happen to be required for a downstream compiler to accept the Slang compiler's output).

This feature is largely motivated by my ongoing work to clean up the Slang language and compiler's rules around checking of initialization behavior.
A common pattern in existing code is that types such as arrays will be initialized using a loop that writes to individual elements:

    int a[10];
    for(int i = 0; i < 10; ++i)
        a[i] = i;
    doSomethingWith(a[someIndex]);

In order for the compiler to validate that the read from `a[someIndex]` is valid (does not access an uninitialized memory location), for an unknown value of `someIndex`, the compiler must confirm that `a` has been fully initialized prior to the call, which in practice means that each of its elements has been fully initialized.
While the above case is straightforward, and some compilers can analyze such loops to confirm initialization status, we do not have the necessary analysis code in Slang at present, and there will always be cases that are too complex for static analysis to handle.

To date we have been dealing with situations like this by simply making our checking for initialization very conservative.
The Slang compiler only diagnoses an issue when it can prove that a read (or read-modify-write) access would *definitely* access an uninitialized memory location, and those diagnostics are currently only warnings, rather than errors.
It is desirable to tighten up this behavior, but doing so means that a lot of existing code written in an idiomatic fashion would start to generate errors.
We need to provide options for users to make such code compatible with any upcoming language version with more strict initialization rules.

For the example above, a conventional way to silence any compiler diagnostic about the access `a[someIndex]` would be to explicitly initialize `a` as part of its declaration, e.g.:

    int a[10] = { 0 };

For small arrays of simple types, such an approach might be fine, but many users of Slang end up working with larger arrays and/or complex types, where the cost of initializing an array up front, only to execute a loop that overwrites all those initial values, is undesirable.

The new `__undefined` literal provides a way for users to indicate to the compiler that they are intentionally leaving a variable in an undefined state at the point of declaration, but that they believe that their code ensures the prior to any read access the variable (or the specific memory location being read) will have been fully initialized to some well-defined value.

Note that the compiler can (and should) still diagnose an problem if it can prove that a read access will access a memory location containing an `__undefined` value, so this feature does not simply give programmers a way to "turn off" the checking logic.
In brief, the desired rules that a new initialization checking pass should implement are as follows:

* If the compiler determines that an access *could* (if executed) read uninitialized memory, it will emit a diagnostic.
* If the compiler can prove that an access *will* (if executed) read memory with `__undefined` contents, it will emit a diagnostic.
* Otherwise, no diagnostic should be emitted for that access.

Note that this change does *not* include a new checking pass that implements the above rules and takes advantage of the information provided by `__undefined` initializers;
this change is just concerned with putting the infrastructure in place.

In addition to adding the `__undefined` literal expression, this change also updates the code for the Slang builtin modules to make use of `__undefined` in places where the (in-progress) validation pass would otherwise emit a diagnostic.
These changes fall into two main categories:

* Places where a loop is used to initialize a variable (or `out` parameter) element-by-element.

* Places where a variable (or `out` parameter) is initialized by code inside a `spirv_asm` block (because these blocks are opaque to the compiler's analysis - see issue #9035).