Provide a proxy slang.dll library on Windows
For customers that have not yet adapted to the new slang-compiler.dll
name, this change provides a slang.dll that forwards all public
functions to slang-compiler.dll.

The intention here is to provide a migration path for existing customers
whose application might have a dependency graph that includes both
slang.dll and slang-compiler.dll. In a case like this, we want all
functions to be handled by the same dll so that we don't end up with
heap corruption or state issues.