LLVM coverage support (#8651)
Add LLVM coverage support with a script to run coverage as well as a new
GitHub CI workflow to run nightly coverage.

The new workflow triggers daily, uses Ubuntu 22.04 and Clang 18 to build
coverage version, and runs slang-test. It produces an HTML report to
GitHub pages as a subdirectory, as well as a gzipped LCOV report as an
artifact.

Fixes #8695
Fixes #8698