Open correct module filename (opening glsl-module)
An earlier change in this patch series used the TAG when generating
the module filename for slang-glsl-module. This is incorrect, since
it often includes more than just the numeric part of the version.
(i.e., it's often the full "git describe")

Adds a new numeric-only define to the slang-tag-version.h header. This
header is part of the public API.