Don't run test if applyBinding fails (#8765)
Small QoL improvement to slang-test; if bindings fail, there is no
reason to run the test, as it will print an error anyway. Because
`applyBinding` bails immediately if any bindings are invalid, valid
bindings can also get skipped, which can then cause crashes and all
kinds of weirdness in the test.

One way to cause these crashes is to add
```slang
//TEST_INPUT: uniform(data=[20 0 0 0]):name=nonexistentbuffer
```
at the beginning of a CPU test; it'll cause the valid buffers after that
to not get bound and their pointers will be left as `nullptr`,
segfaulting in the test. After this PR, the relevant error is printed
without the segfault:
```
error: could not find shader parameter matching 'nonexistentbuffer'
```