Fix undefined behaviour in obfuscate-specialization-naming test (#8764)
This test was reading and writing past the end of the buffers. It
probably didn't cause issues before due to only being run on fairly
resilient drivers, but technically this is UB in Vulkan. Found it while
debugging intermittent test crashes with the LLVM target; this caused
memory corruption there.