Fix `static_assert` incorrectly triggering in dead-branches when `-minimum-slang-optimization` is enabled (#8784)
Fixes #8783

The fix here is to ensure we clean up dead branches after inlining with
SCCP. We do this since we may be given enough information after inlining
to propagate constants up to a point where we can identify a
dead-branch. This then allows us to avoid `static_assert`s in
dead-branches since these branches are eliminated.