Fix a cap issue in __target_switch (#8766)
Close #8771.

The root cause of this issue is that in
`CapabilityDeclReferenceVisitor::visitTargetSwitchStmt`, if there are
two cases that are belong to one target, current implementation just
union the two capability sets, which is wrong.

Instead, the correct operation should be something like this:
```
# compatibleMerge(A, B) =
# ⎧ B   if A = ∅
# ⎪ A   if B = ∅
# ⎪ A   if A ⊆ B
# ⎪ B   if B ⊆ A
# ⎩ ∅   otherwise

where ∅ is empty or undefined set.
```
The only close concept to this operation is `Partial meet` in order
theory. However, we don't choose this name because it's not a common
name and we also extend this operation by handling when A or B is
undefined. So we choose the name `compatibleMerge`

After applying this operation on `target_switch`, we have another
problem in `defaultGetDescriptorFromHandle`, e,g,

```
__target_switch
{
    case spvBindlessTextureNV:
               ...
    case spirv:
          ......
          return __slang_noop_cast<T>(RaytracingAccelerationStructure(__asuint64((uint2)handleValue)));
}
```

in the spirv case, we introduce the capability of `raytracing`. So this
makes the two capability sets as
`spirv + SPV_NV_bindless_texture`
`spirv + raytrcing`
which are disjoint, so according to our definition, merging of those two
set will become undefined.

So, we remove the `raytracing` requirement from
`RaytracingAccelerationStructure`'s 64bit integer constructor, instead
we will just emit the `raytracing` extension from the code generation.
Because this struct is used in both ray tracing and ray query APIs,
however, we don't currently have a way to express a OR requirement. And
this is not the best practice either. The correct way to do it is to
find the common requirement of ray tracing and ray query, which is just
`glsl | spirv`.

So `defaultGetDescriptorFromHandle` on the spirv case won't introduce
`raytracing` anymore.