GLSL: Implement implicit/explicit LOD handling in texture() and variants
The GLSL specification states that texture sampling functions that
require implicit derivatives resort to using MIP=0 when the implicit
derivatives are unavailable [1]. Implement this accordingly.

That is, in texture() and similar:
- Select _Texture.Sample() and similar when the derivatives are
  available
- Select _Texture.SampleLevelZero() and similar when the derivatives
  are unavailable.

To support Slang -> GLSL translation, add
_Texture.SampleLevelZero(). This is the same as _Texture.SampleLevel()
except that the level is constant 0. This is analogous to the
pre-existing _Texture.SampleCmpLevelZero() function. This allows Slang
to emit GLSL texture() with implicit LOD=0 when derivatives are not
available, avoiding corner cases for texture shapes that have
texture() but not the respective textureLod(). Thus, GLSL -> Slang ->
GLSL translation produces valid code for these functions.

The exception to the rule is textureProj(), which will unconditionally
map to textureProj() with LOD=0 when the derivatives are
unavailable. The Slang standard library does not have a _Texture
method for projection lookup, for which the LevelZero() variant could
be added.

The derivatives are always available in the fragment stage. In compute
and raytracing stages, the derivatives are available when the target
has the respective capability. That is:
- GLSL: GL_NV_compute_shader_derivatives:
- SPIR-V: SPV_KHR_compute_shader_derivatives:
- HLSL: SM 6.6

Update the tests accordingly:
- Add tests/bugs/gh-8683-glsl-sample-implicit-lod.slang - direct
  regression test for the issue
- Update tests/glsl-intrinsic/intrinsic-texture.slang - the main workhorse
  for GLSL implicit texture lookup functions
- Update tests/glsl-intrinsic/compute-derivative/intrinsic-derivative-function-in-compute.slang
  for the missing capabilities
- Update tests/language-feature/capability/intrinsic-texture-ignore-capability.slang
  for the missing capabilities

Issue #8683

References:
[1] The OpenGL Shading Language, Version 4.60.8, Section 8.9